Added and modified weapon system, player movement, and input systems.
