+++ Assets/Input/PlayerInput.inputactions.meta
@@ -9,6 +9,6 @@ ScriptedImporter:
   assetBundleVariant: 
   script: {fileID: 11500000, guid: 8404be70184654265930450def6a9037, type: 3}
   generateWrapperCode: 1
-  wrapperCodePath: Assets/Scripts/PlayerInput.cs
+  wrapperCodePath: Assets/Scripts/Player/PlayerInput.cs
   wrapperClassName: 
-  wrapperCodeNamespace: 
+  wrapperCodeNamespace: Player
