+++ Assets/Scripts/Player.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 1ec09bacbc3445839f9a3bceb53122c4
+timeCreated: 1697319874
