+++ Assets/Scripts/Player/PlayerMovement.cs
@@ -1,0 +1,52 @@
+using UnityEngine;
+using UnityEngine.InputSystem;
+
+namespace Player
+{
+    public class PlayerMovement : MonoBehaviour
+    {
+        private PlayerInput input = null;
+        private Rigidbody2D rb = null;
+
+        private Vector2 moveVector = Vector2.zero;
+
+        [SerializeField]
+        private float moveSpeed = 10f;
+
+
+        private void Awake()
+        {
+            input = new PlayerInput();   
+            rb = GetComponent<Rigidbody2D>();
+        }
+
+        private void OnEnable()
+        {
+            input.Enable();
+            input.Player.Movement.performed += OnMovementPerformed;
+            input.Player.Movement.canceled += OnMovementCancelled;
+        }
+
+        private void OnDisable()
+        {
+            input.Disable();
+            input.Player.Movement.performed -= OnMovementPerformed;
+            input.Player.Movement.canceled -= OnMovementCancelled;
+        }
+
+        private void OnMovementPerformed(InputAction.CallbackContext value)
+        {
+            moveVector = value.ReadValue<Vector2>();
+        }
+
+        private void OnMovementCancelled(InputAction.CallbackContext value)
+        {
+            moveVector = Vector2.zero;
+        }
+
+        private void FixedUpdate()
+        {
+            rb.velocity = moveVector * moveSpeed;
+        }
+    }
+}
