+++ Assets/Scripts/Player/PlayerWeaponControl.cs.meta
