+++ Assets/Scripts/Weapons.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 99f34bede1994bdaafd512f8abe3e280
+timeCreated: 1697319912
