+++ Assets/Scripts/Weapons/FireMode.cs
@@ -1,0 +1,9 @@
+namespace Weapons
+{
+    public enum FireMode
+    {
+        Single,
+        Burst,
+        Auto
+    }
+}
