+++ Assets/Scripts/Weapons/FireMode.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 9d9183064e2240e69ef97e4b5fab5bf0
+timeCreated: 1697320995
