+++ Assets/Scripts/Weapons/Pistol.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dd61f434ff0cd9044bcbc1adc2aca226
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
