+++ Assets/Scripts/Weapons/Weapon.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 2f944b6ca52b4ae0947b76b420197594
+timeCreated: 1697319930
