+++ Assets/Scripts/Weapons/WeaponType.cs
@@ -1,0 +1,13 @@
+namespace Weapons
+{
+    public enum WeaponType
+    {
+        Knife,
+        Pistol,
+        Shotgun,
+        SMG,
+        AR,
+        Sniper,
+        MachineGun
+    }
+}
