+++ Assets/Scripts/Weapons/WeaponType.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 9a5a584b48414dbb9856230c444f9c91
+timeCreated: 1697319985
